requirements.txt
PyQt5>=5.15.0
PyMuPDF>=1.18.0
Pillow>=8.0.0